+++ src/freenet/clients/http/WelcomeToadlet.java
@@ -524,7 +524,7 @@ public class WelcomeToadlet extends Toadlet {
 			contentNode.addChild(ctx.getAlertManager().createSummary());
         }
 		
-        if (ctx.getPageMaker().getTheme().fetchKeyBoxAboveBookmarks) {
+        if (node.config.get("fproxy").getBoolean("fetchKeyBoxAboveBookmarks")) {
             this.putFetchKeyBox(ctx, contentNode);
         }
         
@@ -555,7 +555,7 @@ public class WelcomeToadlet extends Toadlet {
         }
 
         // Fetch key box if the theme wants it below the bookmarks.
-        if (!ctx.getPageMaker().getTheme().fetchKeyBoxAboveBookmarks) {
+        if (!node.config.get("fproxy").getBoolean("fetchKeyBoxAboveBookmarks")) {
             this.putFetchKeyBox(ctx, contentNode);
         }