+++ src/freenet/config/Dimension.java
@@ -1,0 +1,6 @@
+package freenet.config;
+
+public enum Dimension {
+
+    NOT, SIZE, DURATION
+}